Decouple storages (#161)
* Decouple storages

* Reorder methods, remove parent call

* Reorder methods 2

* Move creation of empty storages to separate methods

* Reuse manager

* Use trait for manager test

* Compare arrays ignoring order of elements

* Compare arrays ignoring order of elements 2

* Do not create storages on setup

* Ignore compared arrays order

* Ignore compared arrays order 2

* Get children recursive

* Get children recursive

* Move trait back for better diff for review

* Revert

* Reorder methods [skip ci]

* Use ClockMock

* Apply fixes from StyleCI

* Remove final from assignment storage [skip ci]

* Split config and logic

* Add extension to scrutinizer

* Ignore order when comparing arrays

* Ignore order when comparing arrays 2

* Test default value of enable direct permissons

* Move manager arguments logic to a separate method, improve defaults

* Update call for create manager

* Simplify

* Change visibility

* Change visibility 2

* Revert final manager, add interface

* Apply fixes from StyleCI

* Return interface

* Extend interface

* Move PHPDoc to interface

* Apply fixes from StyleCI

* Fix Psalm config and missing import

* Bump PHP version to 8, related fixes and improvements

* Add PHPDoc to interface itself

* Remove PHP 7.4 from workflows

* Add composer-require-checker

* Add rector

* Apply Rector changes (CI)

* Apply fixes from StyleCI

* Sync rector config with the latest changes

* Revert static functions

* Revert rector changes

* Apply fixes from StyleCI

* Revert some rector changes

* Apply fixes from StyleCI

* Update CHANGELOG [skip ci]

* Update .github/workflows/composer-require-checker.yml


* Update .github/workflows/static.yml


* Update src/ManagerInterface.php


* Update src/ManagerInterface.php


* Update src/ManagerInterface.php


* Update src/ManagerInterface.php


* Update src/ManagerInterface.php


* DI + missing types

* Add note [skip ci]

* Fix PHPDoc as well

* Apply fixes from StyleCI

* More strict type hinting for AccessCheckerInterface

* Apply fixes from StyleCI

* Revert recursive children, fix unit tests

* Remove ensureStringUserId

* Apply fixes from StyleCI

* Rename test traits

* Use stable version of yiisoft/access

* Remove empty params

---------
