+++ .scrutinizer.yml
@@ -11,6 +11,8 @@ build:
   environment:
     php:
       version: 8.0.18
+      pecl_extensions:
+        - uopz
       ini:
         xdebug.mode: coverage
