+++ config/di.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+use Yiisoft\Rbac\Manager;
+use Yiisoft\Rbac\ManagerInterface;
+
+return [
+    ManagerInterface::class => Manager::class,
+];
