+++ src/Assignment.php
@@ -10,30 +10,12 @@ namespace Yiisoft\Rbac;
 final class Assignment
 {
     /**
-     * @var string The user ID. This should be a string representing the unique identifier of a user.
-     */
-    private string $userId;
-
-    /**
-     * @var string The role or permission name.
-     */
-    private string $itemName;
-
-    /**
-     * @var int UNIX timestamp representing the assignment creation time.
-     */
-    private int $createdAt;
-
-    /**
      * @param string $userId The user ID. This should be a string representing the unique identifier of a user.
      * @param string $itemName The role or permission name.
      * @param int $createdAt UNIX timestamp representing the assignment creation time.
      */
-    public function __construct(string $userId, string $itemName, int $createdAt)
+    public function __construct(private string $userId, private string $itemName, private int $createdAt)
     {
-        $this->userId = $userId;
-        $this->itemName = $itemName;
-        $this->createdAt = $createdAt;
     }
 
     public function getUserId(): string
