+++ tests/AssignmentTest.php
@@ -11,7 +11,7 @@ final class AssignmentTest extends TestCase
 {
     public function testImmutability(): void
     {
-        $original = new Assignment('42', 'test1', 1642029084);
+        $original = new Assignment('42', 'test1', 1_642_029_084);
         $new = $original->withItemName('test2');
 
         $this->assertNotSame($original, $new);
