+++ tests/Support/RuleWithRequiredName.php
@@ -9,11 +9,8 @@ use Yiisoft\Rbac\RuleInterface;
 
 final class RuleWithRequiredName implements RuleInterface
 {
-    private string $name;
-
-    public function __construct(string $name)
+    public function __construct(private string $name)
     {
-        $this->name = $name;
     }
 
     public function getName(): string
