+++ tests/Support/SimpleRuleFactory.php
@@ -13,16 +13,10 @@ use function array_key_exists;
 final class SimpleRuleFactory implements RuleFactoryInterface
 {
     /**
-     * @psalm-var array<string,RuleInterface>
-     */
-    private array $rules;
-
-    /**
      * @psalm-param array<string,RuleInterface> $rules
      */
-    public function __construct(array $rules = [])
+    public function __construct(private array $rules = [])
     {
-        $this->rules = $rules;
     }
 
     public function create(string $name): RuleInterface
